Split logic in GroupsController into concerns and controllers.
In reference to #33, logic in GroupsController is split into concerns
and controllers.

Concerns are created for features related to:
- Updating individual attributes in Groups (name, grade, project_topic)

Controllers are created for features related to:
- Adding and removing subscribers. (SubscribersController)
- Uploading and deleting project. (ProjectsController)

Additionally:
- Moved add_subscribers view from Groups to Subscribers
- Removed unused ```download_project``` method.
