+++ app/helpers/projects_helper.rb
@@ -1,0 +1,2 @@
+module ProjectsHelper
+end
