+++ app/helpers/subscribers_helper.rb
@@ -1,0 +1,2 @@
+module SubscribersHelper
+end
